update: split highway.py
