+++ cai/client/highway/encoders.py
@@ -1,6 +1,4 @@
-import os
-
-from cai.pb.im.oidb.cmd0x388.cmd0x388_pb2 import D388ReqBody, TryUpImgReq, D388RspBody
+from cai.pb.im.oidb.cmd0x388.cmd0x388_pb2 import D388ReqBody, TryUpImgReq
 
 
 def encode_d388_req(group_code: int, uin: int, md5: bytes, size: int) -> D388ReqBody:
@@ -28,6 +26,3 @@ def encode_d388_req(group_code: int, uin: int, md5: bytes, size: int) -> D388Req
         )]
     )
 
-
-def decode_d388_rsp(data: bytes) -> D388RspBody:
-    return D388RspBody.FromString(data)
